feat: TNL-10136 Tease course enrollment from student model
fix imports and mocking to match new home
