+++ common/djangoapps/student/models/__init__.py
@@ -1,4 +1,5 @@
 '''
 Student models migrated to folder to tease out the course enrollment aspects from Student
 '''
+from .course_enrollment import *
 from .student import *
