+++ common/djangoapps/student/tests/test_enrollment.py
@@ -175,7 +175,7 @@ class EnrollmentTest(UrlResetMixin, ModuleStoreTestCase, OpenEdxEventsTestMixin)
         """Confirm that we send the external updates experiment bucket with the activation signal"""
         with patch('openedx.core.djangoapps.schedules.config.set_up_external_updates_for_enrollment',
                    return_value=value):
-            with patch('common.djangoapps.student.models.student.segment') as mock_segment:
+            with patch('common.djangoapps.student.models.course_enrollment.segment') as mock_segment:
                 CourseEnrollment.enroll(self.user, self.course.id)
 
         assert mock_segment.track.call_count == 1
@@ -183,7 +183,7 @@ class EnrollmentTest(UrlResetMixin, ModuleStoreTestCase, OpenEdxEventsTestMixin)
         assert mock_segment.track.call_args[0][2]['external_course_updates'] == value
 
     def test_enrollment_properties_in_segment_traits(self):
-        with patch('common.djangoapps.student.models.student.segment') as mock_segment:
+        with patch('common.djangoapps.student.models.course_enrollment.segment') as mock_segment:
             enrollment = CourseEnrollment.enroll(self.user, self.course.id)
         assert mock_segment.track.call_count == 1
         assert mock_segment.track.call_args[0][1] == 'edx.course.enrollment.activated'
@@ -192,7 +192,7 @@ class EnrollmentTest(UrlResetMixin, ModuleStoreTestCase, OpenEdxEventsTestMixin)
         assert traits['mode'] == 'audit'
         assert traits['email'] == self.EMAIL
 
-        with patch('common.djangoapps.student.models.student.segment') as mock_segment:
+        with patch('common.djangoapps.student.models.course_enrollment.segment') as mock_segment:
             enrollment.update_enrollment(mode='verified')
         assert mock_segment.track.call_count == 1
         assert mock_segment.track.call_args[0][1] == 'edx.course.enrollment.mode_changed'
