+++ lms/djangoapps/program_enrollments/management/commands/tests/test_send_program_course_nudge_email.py
@@ -81,7 +81,7 @@ class TestSendProgramCourseNudgeEmailCommand(TestCase):
     @ddt.data(
         False, True
     )
-    @patch('common.djangoapps.student.models.student.segment.track')
+    @patch('common.djangoapps.student.models.course_enrollment.segment.track')
     @patch('lms.djangoapps.program_enrollments.management.commands.send_program_course_nudge_email.get_programs')
     @override_settings(FEATURES=dict(ENABLE_ENTERPRISE_INTEGRATION=True))
     def test_email_send(self, add_no_commit, get_programs_mock, mock_track):
@@ -113,7 +113,7 @@ class TestSendProgramCourseNudgeEmailCommand(TestCase):
     @ddt.data(
         False, True
     )
-    @patch('common.djangoapps.student.models.student.segment.track')
+    @patch('common.djangoapps.student.models.course_enrollment.segment.track')
     @patch('lms.djangoapps.program_enrollments.management.commands.send_program_course_nudge_email.get_programs')
     @override_settings(FEATURES=dict(ENABLE_ENTERPRISE_INTEGRATION=True))
     def test_email_no_course_recommendation(self, add_no_commit, get_programs_mock, mock_track):
