+++ openedx/features/enterprise_support/tests/test_signals.py
@@ -133,7 +133,7 @@ class EnterpriseSupportSignals(SharedModuleStoreTestCase):
 
         return enrollment
 
-    @patch('common.djangoapps.student.models.student.CourseEnrollment.is_order_voucher_refundable')
+    @patch('common.djangoapps.student.models.course_enrollment.CourseEnrollment.is_order_voucher_refundable')
     @ddt.data(
         (True, True, 2, True, False),  # test if skip_refund
         (False, True, 20, True, False),  # test refundable time passed
@@ -161,7 +161,7 @@ class EnterpriseSupportSignals(SharedModuleStoreTestCase):
             enrollment.update_enrollment(is_active=False, skip_refund=skip_refund)
             assert mock_ecommerce_api_client.called == api_called
 
-    @patch('common.djangoapps.student.models.student.CourseEnrollment.is_order_voucher_refundable')
+    @patch('common.djangoapps.student.models.course_enrollment.CourseEnrollment.is_order_voucher_refundable')
     @ddt.data(
         (HttpClientError, 'INFO'),
         (HttpServerError, 'ERROR'),
